Revert "Lsf annotation"
