+++ src/redux/actions/api/Tasks/GetNextTask.js
@@ -6,17 +6,16 @@ import ENDPOINTS from "../../../../config/apiendpoint";
 import constants from "../../../constants";
 
 export default class GetNextTaskAPI extends API {
-  constructor(projectId,projectObj ,timeout = 2000) {
+  constructor(projectId, taskId, mode="annotation", timeout = 2000) {
     super("POST", timeout, false);
     this.projectId = projectId;
     this.labellingMode = localStorage.getItem("labellingMode");
     this.searchFilters = JSON.parse(localStorage.getItem("searchFilters"));
-    this.projectObj = projectObj;
     this.type = constants.GET_NEXT_TASK;
-    this.endpoint = `${super.apiEndPointAuto()}${ENDPOINTS.getProjects}${projectId}/next/`;
-    if (localStorage.getItem("labelAll") ) {
+    this.endpoint = `${super.apiEndPointAuto()}${ENDPOINTS.getProjects}${projectId}/next/?${taskId ? 'current_task_id=' + taskId : ''}${this.labellingMode ? '&task_status=' + this.labellingMode : ''}${mode === "review" ? '&mode=review' : ''}`;
+    if (localStorage.getItem("labelAll")) {
       Object.keys(this.searchFilters).forEach(key => {
-        this.endpoint += `?${key}=${this.searchFilters[key]}`;
+        this.endpoint += `&${key}=${this.searchFilters[key]}`;
       });
     }
   }
@@ -33,7 +32,7 @@ export default class GetNextTaskAPI extends API {
   }
 
   getBody() {
-    return this.projectObj;
+    return { id: this.projectId };
   }
 
   getHeaders() {
