+++ src/ui/pages/component/Project/TaskTable.jsx
@@ -151,23 +151,15 @@ const TaskTable = (props) => {
         }
     }
 
-  
-
     const labelAllTasks = () => {
-
         let search_filters = Object.keys(selectedFilters).filter(key => key.startsWith("search_")).reduce((acc, curr) => {
             acc[curr] = selectedFilters[curr];
             return acc;
         }, {});
-        localStorage.setItem("labellingMode", props.type === "annotation" ? selectedFilters.annotation_status:selectedFilters.review_status);
+        localStorage.setItem("labellingMode", selectedFilters.task_status);
         localStorage.setItem("searchFilters", JSON.stringify(search_filters));
         localStorage.setItem("labelAll", true);
-       
-        const datavalue= {
-            annotation_status:selectedFilters?.annotation_status,
-            ...( props.type === "review" && {mode: "review", annotation_status:selectedFilters?.review_status})
-        }
-        const getNextTaskObj = new GetNextTaskAPI( id,datavalue,null, props.type);
+        const getNextTaskObj = new GetNextTaskAPI(id, null, props.type);
         dispatch(APITransport(getNextTaskObj));
         setLabellingStarted(true);
     };
