+++ src/ui/pages/container/Label-Studio/LSF.jsx
@@ -1,20 +1,12 @@
-import PropTypes from "prop-types";
+import PropTypes from 'prop-types'
 import React, { useState, useEffect, useRef } from "react";
 import LabelStudio from "@heartexlabs/label-studio";
-import {
-  Tooltip,
-  Button,
-  Alert,
-  Card,
-  TextField,
-  Box,
-  Grid,
-} from "@mui/material";
-import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
-import ArrowRightIcon from "@mui/icons-material/ArrowRight";
+import { Tooltip, Button, Alert, Card, TextField, Box, Grid } from "@mui/material";
+import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
+import ArrowRightIcon from '@mui/icons-material/ArrowRight';
 import CustomizedSnackbars from "../../component/common/Snackbar";
-import generateLabelConfig from "../../../../utils/LabelConfig/ConversationTranslation";
+import generateLabelConfig from '../../../../utils/LabelConfig/ConversationTranslation';
 
 import {
   getProjectsandTasks,
@@ -22,36 +14,28 @@ import {
   getNextProject,
   patchAnnotation,
   deleteAnnotation,
-  fetchAnnotation,
+  fetchAnnotation
 } from "../../../../redux/actions/api/LSFAPI/LSFAPI";
 import GetProjectDetailsAPI from "../../../../redux/actions/api/ProjectDetails/GetProjectDetails";
-import APITransport from "../../../../redux/actions/apitransport/apitransport";
+import APITransport from '../../../../redux/actions/apitransport/apitransport';
 
 import { useParams, useNavigate } from "react-router-dom";
 import useFullPageLoader from "../../../../hooks/useFullPageLoader";
-import { snakeToTitleCase } from "../../../../utils/utils";
+import { snakeToTitleCase } from '../../../../utils/utils';
 
-import styles from "./lsf.module.css";
-import "./lsf.css";
-import { useDispatch, useSelector } from "react-redux";
-import { translate } from "../../../../config/localisation";
+import styles from './lsf.module.css'
+import "./lsf.css"
+import { useDispatch, useSelector } from 'react-redux';
+import { translate } from '../../../../config/localisation';
 
 //used just in postAnnotation to support draft status update.
 
-const LabelStudioWrapper = ({
-  annotationNotesRef,
-  loader,
-  showLoader,
-  hideLoader,
-  resetNotes,
-}) => {
+const LabelStudioWrapper = ({annotationNotesRef, loader, showLoader, hideLoader, resetNotes}) => {
   // we need a reference to a DOM node here so LSF knows where to render
   const rootRef = useRef();
   const dispatch = useDispatch();
-  const ProjectDetails = useSelector((state) => state.getProjectDetails.data);
-  const annotation_status = useRef(
-    ProjectDetails.enable_task_reviews ? "labeled" : "accepted"
-  );
+  const ProjectDetails = useSelector(state => state.getProjectDetails.data);
+  const annotation_status = useRef(ProjectDetails.enable_task_reviews ? "labeled": "accepted");
   // this reference will be populated when LSF initialized and can be used somewhere else
   const lsfRef = useRef();
   const navigate = useNavigate();
@@ -63,43 +47,40 @@ const LabelStudioWrapper = ({
   });
   const [taskData, setTaskData] = useState(undefined);
   const { projectId, taskId } = useParams();
-  const userData = useSelector((state) => state.fetchLoggedInUserData.data);
+  const userData = useSelector(state=>state.fetchLoggedInUserData.data)
   let loaded = useRef();
 
   console.log("projectId, taskId", projectId, taskId);
   // debugger
 
   useEffect(() => {
-    localStorage.setItem(
-      "labelStudio:settings",
-      JSON.stringify({
-        bottomSidePanel: true,
-        continuousLabeling: false,
-        enableAutoSave: false,
-        enableHotkeys: true,
-        enableLabelTooltips: true,
-        enablePanelHotkeys: true,
-        enableTooltips: false,
-        fullscreen: false,
-        imageFullSize: false,
-        selectAfterCreate: false,
-        showAnnotationsPanel: true,
-        showLabels: false,
-        showLineNumbers: false,
-        showPredictionsPanel: true,
-        sidePanelMode: "SIDEPANEL_MODE_REGIONS",
-      })
-    );
-  }, []);
+    localStorage.setItem("labelStudio:settings", JSON.stringify({
+      bottomSidePanel: true,
+      continuousLabeling: false,
+      enableAutoSave: false,
+      enableHotkeys: true,
+      enableLabelTooltips: true,
+      enablePanelHotkeys: true,
+      enableTooltips: false,
+      fullscreen: false,
+      imageFullSize: false,
+      selectAfterCreate: false,
+      showAnnotationsPanel: true,
+      showLabels: false,
+      showLineNumbers: false,
+      showPredictionsPanel: true,
+      sidePanelMode: "SIDEPANEL_MODE_REGIONS"
+    }))
+  }, [])
 
   const tasksComplete = (id) => {
     if (id) {
-      resetNotes();
+      resetNotes()
       // navigate(`/projects/${projectId}/task/${id}`, {replace: true});
       navigate(`/projects/${projectId}/task/${id}`);
     } else {
       // navigate(-1);
-      resetNotes();
+      resetNotes()
       setSnackbarInfo({
         open: true,
         message: "No more tasks to label",
@@ -111,7 +92,7 @@ const LabelStudioWrapper = ({
         window.location.reload();
       }, 1000);
     }
-  };
+  }
 
   function LSFRoot(
     rootRef,
@@ -122,7 +103,7 @@ const LabelStudioWrapper = ({
     labelConfig,
     annotations,
     predictions,
-    annotationNotesRef
+    annotationNotesRef,
   ) {
     let load_time;
     let interfaces = [];
@@ -157,13 +138,11 @@ const LabelStudioWrapper = ({
         "update",
         "submit",
         "skip",
-        ...(taskData?.annotation_users?.some((user) => user === userData.id)
-          ? ["controls"]
-          : []),
+        ...taskData?.annotation_users?.some((user) => user === userData.id) ? ["controls"] : [],
         "infobar",
         "topbar",
         "instruction",
-        "side-column",
+         "side-column",
         "annotations:history",
         "annotations:tabs",
         "annotations:menu",
@@ -202,15 +181,8 @@ const LabelStudioWrapper = ({
         },
 
         onLabelStudioLoad: function (ls) {
-          annotation_status.current = ProjectDetails.enable_task_reviews
-            ? "labeled"
-            : "accepted";
-          console.log(
-            "annotation_status",
-            annotation_status.current,
-            "test",
-            ProjectDetails
-          );
+          annotation_status.current = ProjectDetails.enable_task_reviews ? "labeled": "accepted";
+          console.log("annotation_status", annotation_status.current, "test", ProjectDetails);
           if (annotations.length === 0) {
             var c = ls.annotationStore.addAnnotation({
               userGenerate: true,
@@ -236,26 +208,25 @@ const LabelStudioWrapper = ({
                   hideLoader();
                   // window.location.href = `/projects/${projectId}/task/${res.id}`;
                   tasksComplete(res?.id || null);
-                });
+                })
               else {
                 hideLoader();
                 window.location.reload();
               }
-            });
-          } else
-            setSnackbarInfo({
-              open: true,
-              message: "Task is frozen",
-              variant: "error",
-            });
+            })
+          }
+          else
+          setSnackbarInfo({
+            open: true,
+            message: "Task is frozen",
+            variant: "error",
+          });
         },
 
         onSkipTask: function () {
-          //   message.warning('Notes will not be saved for skipped tasks!');
-          let annotation = annotations.find(
-            (annotation) => !annotation.parentAnnotation
-          );
-          console.log("onSkip", annotation);
+        //   message.warning('Notes will not be saved for skipped tasks!');
+          let annotation = annotations.find((annotation) => !annotation.parentAnnotation);
+          console.log("onSkip", annotation)
           if (annotation) {
             showLoader();
             patchAnnotation(
@@ -277,17 +248,13 @@ const LabelStudioWrapper = ({
         onUpdateAnnotation: function (ls, annotation) {
           if (taskData.annotation_status !== "freezed") {
             for (let i = 0; i < annotations.length; i++) {
-              if (
-                !annotations[i].result?.length ||
-                annotation.serializeAnnotation()[0].id ===
-                  annotations[i].result[0].id
-              ) {
+              if (!annotations[i].result?.length || annotation.serializeAnnotation()[0].id === annotations[i].result[0].id) {
                 showLoader();
-                let temp = annotation.serializeAnnotation();
+                let temp = annotation.serializeAnnotation()
 
                 for (let i = 0; i < temp.length; i++) {
                   if (temp[i].value.text) {
-                    temp[i].value.text = [temp[i].value.text[0]];
+                    temp[i].value.text = [temp[i].value.text[0]]
                   }
                 }
                 patchAnnotation(
@@ -297,41 +264,41 @@ const LabelStudioWrapper = ({
                   annotations[i].lead_time,
                   annotation_status.current,
                   annotationNotesRef.current.value
-                ).then(() => {
-                  if (localStorage.getItem("labelAll"))
-                    getNextProject(projectId, taskData.id).then((res) => {
+                  ).then(() => {
+                    if (localStorage.getItem("labelAll"))
+                      getNextProject(projectId, taskData.id).then((res) => {
+                        hideLoader();
+                        tasksComplete(res?.id || null);
+                      })
+                    else{
                       hideLoader();
-                      tasksComplete(res?.id || null);
-                    });
-                  else {
-                    hideLoader();
-                    window.location.reload();
-                  }
-                });
+                      window.location.reload();
+                    }
+                  });
               }
             }
-          } else
-            setSnackbarInfo({
-              open: true,
-              message: "Task is frozen",
-              variant: "error",
-            });
+          } 
+          else
+          setSnackbarInfo({
+            open: true,
+            message: "Task is frozen",
+            variant: "error",
+          });
         },
 
         onDeleteAnnotation: function (ls, annotation) {
           for (let i = 0; i < annotations.length; i++) {
-            if (
-              annotation.serializeAnnotation()[0].id ===
-              annotations[i].result[0].id
-            ) {
-              deleteAnnotation(annotations[i].id);
+            if (annotation.serializeAnnotation()[0].id === annotations[i].result[0].id) {
+              deleteAnnotation(
+                annotations[i].id
+              );
               var c = ls.annotationStore.addAnnotation({
                 userGenerate: true,
               });
               ls.annotationStore.selectAnnotation(c.id);
             }
           }
-        },
+        }
       });
     }
   }
@@ -339,12 +306,14 @@ const LabelStudioWrapper = ({
   // we're running an effect on component mount and rendering LSF inside rootRef node
   localStorage.setItem("TaskData", JSON.stringify(taskData));
   useEffect(() => {
-    if (localStorage.getItem("rtl") === "true") {
-      var style = document.createElement("style");
-      style.innerHTML = "input, textarea { direction: RTL; }";
+    if (localStorage.getItem('rtl') === "true") {
+      var style = document.createElement('style');
+      style.innerHTML = 'input, textarea { direction: RTL; }'
       document.head.appendChild(style);
     }
-    if (userData?.id && loaded.current !== taskId) {
+    if (
+      userData?.id && loaded.current !== taskId
+    ) {
       if (Object.keys(ProjectDetails).length === 0) {
         const projectObj = new GetProjectDetailsAPI(projectId);
         dispatch(APITransport(projectObj));
@@ -353,17 +322,8 @@ const LabelStudioWrapper = ({
         getProjectsandTasks(projectId, taskId).then(
           ([labelConfig, taskData, annotations, predictions]) => {
             // both have loaded!
-            console.log("[labelConfig, taskData, annotations, predictions]", [
-              labelConfig,
-              taskData,
-              annotations,
-              predictions,
-            ]);
-            let tempLabelConfig =
-              labelConfig.project_type === "ConversationTranslation" ||
-              labelConfig.project_type === "ConversationTranslationEditing"
-                ? generateLabelConfig(taskData.data)
-                : labelConfig.label_config;
+            console.log("[labelConfig, taskData, annotations, predictions]", [labelConfig, taskData, annotations, predictions]);
+            let tempLabelConfig = labelConfig.project_type === "ConversationTranslation" || labelConfig.project_type === "ConversationTranslationEditing" ? generateLabelConfig(taskData.data) : labelConfig.label_config;
             setLabelConfig(tempLabelConfig);
             setTaskData(taskData);
             LSFRoot(
@@ -391,16 +351,16 @@ const LabelStudioWrapper = ({
   const handleDraftAnnotationClick = async () => {
     annotation_status.current = "draft";
     lsfRef.current.store.submitAnnotation();
-  };
+  }
 
   const onNextAnnotation = async () => {
     showLoader();
     getNextProject(projectId, taskId).then((res) => {
       hideLoader();
       // window.location.href = `/projects/${projectId}/task/${res.id}`;
-      tasksComplete(res?.id || null);
+     tasksComplete(res?.id || null);
     });
-  };
+  }
 
   const renderSnackBar = () => {
     return (
@@ -419,77 +379,48 @@ const LabelStudioWrapper = ({
 
   return (
     <div>
-      {!loader && (
-        <div
-          style={{ display: "flex", justifyContent: "space-between" }}
-          className="lsf-controls"
-        >
-          <Grid container spacing={0}>
-            <Grid container spacing={0} sx={{ justifyContent: "end" }}>
-            
-              <Grid item>
-                {localStorage.getItem("labelAll") === "true" ? (
-                  <>
-                    <Tooltip title="Go to next task">
-                      <Button
-                        value="Next"
-                        type="default"
-                        onClick={onNextAnnotation}
-                        style={{
-                          minWidth: "160px",
-                          border: "1px solid #e6e6e6",
-                          color: "#09f",
-                          pt: 3,
-                          pb: 3,
-                          borderBottom: "None",
-                        }}
-                        className="lsf-button"
-                      >
-                        Next
-                      </Button>
-                    </Tooltip>
-
-                  
-                  </>
-                ) : (
-                  <div style={{ minWidth: "160px" }} />
-                )}
-              </Grid>
-              <Grid item>
-                {/* {taskData?.annotation_users?.some(
-                  (user) => user === userData.id
-                ) && ( */}
-                  <Tooltip title="Save task for later">
-                    <Button
-                      value="Draft"
-                      type="default"
-                      onClick={handleDraftAnnotationClick}
-                      style={{
-                        minWidth: "160px",
-                        border: "1px solid #e6e6e6",
-                        color: "#e80",
-                        pt: 3,
-                        pb: 3,
-                        borderBottom: "None",
-                      }}
-                      className="lsf-button"
-                    >
-                      Draft
-                    </Button>
-                  </Tooltip>
-                 
-                {/* )} */}
-              </Grid>
-                   
-            </Grid>
+      {!loader && <div style={{ display: "flex", justifyContent: "space-between" }} className="lsf-controls">
+        <div/>
+        <Grid container spacing={0}>
+        {/* <Grid container spacing={0} sx={{ justifyContent: "end" }}> */}
+          <Grid item>
+          {taskData?.annotation_users?.some((user) => user === userData.id) && <Tooltip title="Save task for later">
+            <Button
+              value="Draft"
+              type="default"
+              onClick={handleDraftAnnotationClick}
+              style={{minWidth: "160px", border:"1px solid #e6e6e6", color: "#e80", pt: 3, pb: 3, borderBottom: "None"}}
+              className="lsf-button"
+            >
+              Draft
+            </Button>
+          </Tooltip>}
           </Grid>
-          <div />
-        </div>
-      )}
-      <Box sx={{ border: "1px solid rgb(224 224 224)" }}>
+          <Grid item>
+          {localStorage.getItem("labelAll") !== "true" ? (
+            <Tooltip title="Go to next task">
+              <Button
+                value="Next"
+                type="default"
+                onClick={onNextAnnotation}
+                style={{minWidth: "160px", border:"1px solid #e6e6e6", color: "#09f", pt: 3, pb: 3, borderBottom: "None",}}
+                className="lsf-button"
+              >
+                Next
+              </Button>
+            </Tooltip>
+          ) : (
+            <div style={{minWidth: "160px"}}/>
+          )}
+          </Grid>
+        </Grid>
+      </div>}
+      <Box
+        sx={{border : "1px solid rgb(224 224 224)"}}
+      >
         <div className="label-studio-root" ref={rootRef}></div>
       </Box>
-
+      
       {loader}
       {renderSnackBar()}
     </div>
@@ -500,119 +431,103 @@ export default function LSF() {
   const [showNotes, setShowNotes] = useState(false);
   const annotationNotesRef = useRef(null);
   const reviewNotesRef = useRef(null);
-  const { taskId } = useParams();
+  const {taskId} = useParams()
   // const [notesValue, setNotesValue] = useState('');
   const { projectId } = useParams();
   const navigate = useNavigate();
   const [loader, showLoader, hideLoader] = useFullPageLoader();
-
+  
   const handleCollapseClick = () => {
     setShowNotes(!showNotes);
-  };
+  }
 
-  useEffect(() => {
-    fetchAnnotation(taskId).then((data) => {
-      if (data && Array.isArray(data) && data.length > 0) {
-        annotationNotesRef.current.value = data[0].annotation_notes ?? "";
-        reviewNotesRef.current.value = data[0].review_notes ?? "";
+  useEffect(()=>{
+    fetchAnnotation(taskId).then((data)=>{
+      if(data && Array.isArray(data) && data.length > 0) {
+        annotationNotesRef.current.value = data[0].annotation_notes?? '';
+        reviewNotesRef.current.value = data[0].review_notes?? '';
       }
-    });
+    })
   }, [taskId]);
 
   const resetNotes = () => {
     setShowNotes(false);
     annotationNotesRef.current.value = "";
     reviewNotesRef.current.value = "";
-  };
+  }
 
-  useEffect(() => {
+  useEffect(()=>{
     resetNotes();
   }, [taskId]);
-
+  
   return (
     <div style={{ maxHeight: "100%", maxWidth: "100%", margin: "auto" }}>
-      {!loader && (
-        <Button
-          value="Back to Project"
-          startIcon={<ArrowBackIcon />}
-          variant="contained"
-          color="primary"
-          onClick={() => {
-            localStorage.removeItem("labelAll");
-            navigate(`/projects/${projectId}`);
-            //window.location.replace(`/#/projects/${projectId}`);
-            //window.location.reload();
-          }}
-        >
-          Back to Project
-        </Button>
-      )}
-      <Card
-        sx={{
-          minHeight: 500,
-          padding: 5,
-          mt: 3,
-          pt: 3,
+      {!loader && <Button
+        value="Back to Project"
+        startIcon={<  ArrowBackIcon />}
+        variant="contained" 
+        color="primary"
+        onClick={() => {
+          localStorage.removeItem("labelAll");
+          navigate(`/projects/${projectId}`);
+          //window.location.replace(`/#/projects/${projectId}`);
+          //window.location.reload();
         }}
       >
-        {!loader && (
-          <Button
-            endIcon={showNotes ? <ArrowRightIcon /> : <ArrowDropDownIcon />}
-            variant="contained"
-            color={reviewNotesRef.current?.value !== "" ? "success" : "primary"}
-            onClick={handleCollapseClick}
-            style={{ marginBottom: "20px" }}
-          >
-            Notes {reviewNotesRef.current?.value !== "" && "*"}
-          </Button>
-        )}
-        <div
-          className={styles.collapse}
-          style={{
-            display: showNotes ? "block" : "none",
-            paddingBottom: "16px",
-          }}
+        Back to Project
+      </Button>}
+      <Card
+        sx={{
+            minHeight: 500,
+            padding: 5,
+            mt: 3,
+            pt: 3,
+        }}
+      >
+        {!loader && <Button 
+          endIcon={showNotes ? <ArrowRightIcon /> : <ArrowDropDownIcon />}
+          variant="contained"
+          color={reviewNotesRef.current?.value !== "" ? "success" : "primary"}
+          onClick={handleCollapseClick}
+          style={{marginBottom:'20px'}}
         >
-          <Alert severity="warning" showIcon style={{ marginBottom: "1%" }}>
-            {translate("alert.notes")}
+          Notes {reviewNotesRef.current?.value !== "" && "*"}
+        </Button>}
+        <div className={styles.collapse} style={{display: showNotes? "block" : "none",paddingBottom: "16px"}}>
+          <Alert severity="warning" showIcon style={{marginBottom: '1%'}}>
+              {translate("alert.notes")}
           </Alert>
           <TextField
-            multiline
-            placeholder="Place your remarks here ..."
+            multiline 
+            placeholder="Place your remarks here ..." 
             label="Annotation Notes"
-            // value={notesValue}
-            // onChange={event=>setNotesValue(event.target.value)}
+            // value={notesValue} 
+            // onChange={event=>setNotesValue(event.target.value)} 
             inputRef={annotationNotesRef}
             rows={2}
             maxRows={4}
             inputProps={{
-              style: { fontSize: "1rem" },
+              style: {fontSize: "1rem",},
             }}
-            style={{ width: "99%" }}
+            style={{width: '99%'}}
           />
           <TextField
-            multiline
-            placeholder="Place your remarks here ..."
+            multiline 
+            placeholder="Place your remarks here ..." 
             label="Review Notes"
-            // value={notesValue}
-            // onChange={event=>setNotesValue(event.target.value)}
+            // value={notesValue} 
+            // onChange={event=>setNotesValue(event.target.value)} 
             inputRef={reviewNotesRef}
             rows={2}
             maxRows={4}
             inputProps={{
-              style: { fontSize: "1rem" },
+              style: {fontSize: "1rem",},
               readOnly: true,
             }}
-            style={{ width: "99%", marginTop: "1%" }}
+            style={{width: '99%', marginTop: '1%'}}
           />
         </div>
-        <LabelStudioWrapper
-          resetNotes={() => resetNotes()}
-          annotationNotesRef={annotationNotesRef}
-          loader={loader}
-          showLoader={showLoader}
-          hideLoader={hideLoader}
-        />
+        <LabelStudioWrapper resetNotes={()=>resetNotes()} annotationNotesRef={annotationNotesRef} loader={loader} showLoader={showLoader} hideLoader={hideLoader}/>
       </Card>
     </div>
   );
