+++ src/web.route.js
@@ -37,7 +37,6 @@ import SignUp from "./ui/pages/container/UserManagement/SignUp";
 import AutomateDatasets from "./ui/pages/container/Dataset/AutomateDatasets";
 import ProgressList from "./ui/pages/container/Progress/ProgressList"
 import browserhistory from "./web.history";
-import AllTaskLSF from "./ui/pages/container/Label-Studio/AllTaskLSF";
 
 
 const App = () => {
@@ -173,13 +172,6 @@ const App = () => {
           path="projects/:projectId/review/:taskId"
           element={ProtectedRouteWrapper(
             <Layout component={<ReviewLSF />} />
-          )}
-        />
-         <Route
-          path="projects/:projectId/Alltask/:taskId"
-          element={ProtectedRouteWrapper(
-            <Layout component={<AllTaskLSF />} />
-            // <Layout component={<AnnotateTask />} />
           )}
         />
